chore(dragging): Rename core/interfaces/i_draggable.ts
Rename core/interfaces/i_draggable.ts to
core/interfaces/i_draggable.old.ts to make room for new
IDraggable.  Do not rename actual interface as it's not yet
clear that it will be necessary for both to coexist as
imports in the same file.
