+++ core/delete_area.ts
@@ -15,7 +15,7 @@
 import {BlockSvg} from './block_svg.js';
 import {DragTarget} from './drag_target.js';
 import type {IDeleteArea} from './interfaces/i_delete_area.js';
-import type {IDraggable} from './interfaces/i_draggable.js';
+import type {IDraggable} from './interfaces/i_draggable.old.js';
 
 /**
  * Abstract class for a component that can delete a block or bubble that is
