+++ core/interfaces/i_bubble.ts
@@ -8,7 +8,7 @@ import type {Coordinate} from '../utils/coordinate.js';
 // Former goog.module ID: Blockly.IBubble
 
 import type {IContextMenu} from './i_contextmenu.js';
-import type {IDraggable} from './i_draggable.js';
+import type {IDraggable} from './i_draggable.old.js';
 
 /**
  * A bubble interface.
