+++ core/interfaces/i_delete_area.ts
@@ -7,7 +7,7 @@
 // Former goog.module ID: Blockly.IDeleteArea
 
 import type {IDragTarget} from './i_drag_target.js';
-import type {IDraggable} from './i_draggable.js';
+import type {IDraggable} from './i_draggable.old.js';
 
 /**
  * Interface for a component that can delete a block or bubble that is dropped
