+++ core/toolbox/toolbox.ts
@@ -24,7 +24,7 @@ import {DeleteArea} from '../delete_area.js';
 import * as eventUtils from '../events/utils.js';
 import type {IAutoHideable} from '../interfaces/i_autohideable.js';
 import type {ICollapsibleToolboxItem} from '../interfaces/i_collapsible_toolbox_item.js';
-import type {IDraggable} from '../interfaces/i_draggable.js';
+import type {IDraggable} from '../interfaces/i_draggable.old.js';
 import type {IFlyout} from '../interfaces/i_flyout.js';
 import type {IKeyboardAccessible} from '../interfaces/i_keyboard_accessible.js';
 import type {ISelectableToolboxItem} from '../interfaces/i_selectable_toolbox_item.js';
